Organize tests into descriptive modules
